+++ instagram/comment/models.py
@@ -1,0 +1,4 @@
+from database.core import Base
+from sqlalchemy import Column, Integer, String
+from sqlalchemy.ext.hybrid import hybrid_property
+from sqlalchemy.orm import relationship
