+++ instagram/post/service.py
@@ -3,7 +3,7 @@ from typing import List, Optional, Type
 from instagram.user.models import InstagramUser
 from sqlalchemy.orm import Session
 
-from .models import Comment, Image, Like, LikeableEntity, Post, PostCreate
+from .models import Comment, Image, LikeableEntity, Post, PostCreate
 
 
 def add_user_meta(post: Post, current_user: InstagramUser) -> None:
@@ -76,28 +76,5 @@ def add_comment(
 def uncomment(db: Session, current_post: Post, comment_id: int):
     db.query(Comment).filter(Comment.post_id == current_post.id).filter(
         Comment.id == comment_id
-    ).delete()
-    db.commit()
-
-
-def like_post_or_comment(
-    db: Session, current_user: InstagramUser, current_entity: LikeableEntity
-) -> None:
-    # a user can only like a post/comment once
-    if (
-        db.query(InstagramUser)
-        .filter(InstagramUser.likes.any(entity_id=current_entity.id))
-        .count()
-        == 0
-    ):
-        db.add(Like(entity_id=current_entity.id, user_id=current_user.id))
-        db.commit()
-
-
-def dislike_post_or_comment(
-    db: Session, current_entity: LikeableEntity, current_user: InstagramUser
-) -> None:
-    db.query(Like).filter(Like.entity_id == current_entity.id).filter(
-        Like.user_id == current_user.id
     ).delete()
     db.commit()
